+++ Assets/Programing/PJS/Prefabs/MosterPoints.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ddf16007a80461246b147cee7eab7c64
+guid: ae3e85a456fc027488ee4565ec992cb5
 PrefabImporter:
   externalObjects: {}
   userData:
