--- Assets/Programing/PJS/Prefabs/Test/BlueGollem_elite.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2cf5aeb9240cc2f4a909f1f0c7fb57e2
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
