--- Assets/Programing/PJS/Prefabs/Test/ClearBox.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2ea9aba3c8aff3f499cac2979860905c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
