+++ Assets/Programing/PJS/Scripts/CreateStageMonster.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f45dffaee48542f4ca9efd9c46accb76
+guid: a47b9fffd8f3ba544addeca9f5ab4b1a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
