+++ Assets/Programing/PJS/Scripts/MonsterData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 273d31b7592c4fa4b9d15d2767355ff3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
